switch to smaller view and remove channel/nicks for small screens, instead of just crashing
